Bug Fix: Truncate unrequired beams from MS files.
When reading MS files and if N_station is set (LOFAR), the visibility
matrix cannot be formed due to a mismatch between the size of S and that
of beam_id (due to N_station).
The solution is to filter the entries in S at the cost of slowing down
the file reads.